restructring bug fix
