+++ src/wallets/index.js
@@ -1,0 +1,5 @@
+export * from './wallet-list';
+export * from './plug';
+export * from './stoic';
+export * from './bitfinity';
+export * from './dfinity';
